refactor: share the JIM SQL Connector's column catalogue read between discovery and export (#170)
Schema discovery parsed the dialect's ColumnsCommandText rows itself. Export is
about to need the same answer, so the parsing moves to SqlCatalogueReader and
discovery calls it. Two parsers of the same rows would eventually disagree about
what a column is.
